requirements.txt
asgiref==3.3.4
astroid==2.5.2
attrs==20.3.0
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
flake8==3.9.0
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
pyjwt==1.7.1
